test(dcode): split image contracts by boundary
Extract credential-boundary coverage and shared wrapper fixtures so each
focused suite stays below the repository test-file ceiling. Update source
comments to follow the moved canonical secret corpus.
